test: applied changes from code review
